address code review; reorganize
