+++ examples/demo-rollup/tests/bank.rs
@@ -4,32 +4,11 @@ use std::net::SocketAddr;
 use borsh::BorshSerialize;
 use demo_stf::app::DefaultPrivateKey;
 use demo_stf::runtime::RuntimeCall;
-use jsonrpsee::core::client::ClientT;
-use jsonrpsee::http_client::{HttpClient, HttpClientBuilder};
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::transaction::Transaction;
 use sov_modules_api::{PrivateKey, Spec};
+use sov_sequencer::utils::SimpleClient;
 use test_helpers::start_rollup;
-
-struct TestClient {
-    client: HttpClient,
-}
-
-impl TestClient {
-    pub fn new(endpoint: &str) -> Self {
-        let client = HttpClientBuilder::default().build(endpoint).unwrap();
-        Self { client }
-    }
-    pub async fn send_transaction(
-        &self,
-        tx: Transaction<DefaultContext>,
-    ) -> Result<(), anyhow::Error> {
-        let batch = vec![tx.try_to_vec()?];
-        let response: String = self.client.request("sequencer_publishBatch", batch).await?;
-        println!("response: {:?}", response);
-        Ok(())
-    }
-}
 
 async fn send_test_create_token_tx(rpc_address: SocketAddr) -> Result<(), anyhow::Error> {
     let key = DefaultPrivateKey::generate();
@@ -44,13 +23,13 @@ async fn send_test_create_token_tx(rpc_address: SocketAddr) -> Result<(), anyhow
     });
     let tx = Transaction::<DefaultContext>::new_signed_tx(&key, msg.try_to_vec().unwrap(), 0);
 
-    let client = TestClient::new(&format!("http://localhost:{}", rpc_address.port()));
+    let client = SimpleClient::new(&format!("http://localhost:{}", rpc_address.port()));
 
     client.send_transaction(tx).await
 }
 
 #[tokio::test]
-async fn tx_tests() -> Result<(), anyhow::Error> {
+async fn bank_tx_tests() -> Result<(), anyhow::Error> {
     let (port_tx, port_rx) = tokio::sync::oneshot::channel();
 
     let rollup_task = tokio::spawn(async {
