+++ examples/demo-rollup/tests/evm_integration_tests.rs
@@ -132,7 +132,7 @@ async fn send_tx_test_to_eth(rpc_address: SocketAddr) -> Result<(), Box<dyn std:
 }
 
 #[tokio::test]
-async fn tx_tests() -> Result<(), anyhow::Error> {
+async fn evm_tx_tests() -> Result<(), anyhow::Error> {
     let (port_tx, port_rx) = tokio::sync::oneshot::channel();
 
     let rollup_task = tokio::spawn(async {
