+++ full-node/sov-sequencer/Cargo.toml
@@ -19,9 +19,7 @@ hex = { workspace = true }
 jsonrpsee = { workspace = true, features = ["http-client", "server"] }
 serde = { workspace = true, features = ["derive"] }
 tracing = { workspace = true }
-
 sov-rollup-interface = { path = "../../rollup-interface" }
-sov-modules-api = { path = "../../module-system/sov-modules-api", features = ["native"] }
 
 [dev-dependencies]
 async-trait = { workspace = true }
